feat: dto 반환타입 변경하여 600ms -> 300ms 로 개선
[#613-1]

+++ backend/src/main/java/com/carffeine/carffeine/station/domain/station/Station.java
@@ -80,12 +80,12 @@ public class Station extends BaseEntity {
     @OneToMany(mappedBy = "station", fetch = FetchType.LAZY, cascade = CascadeType.ALL, orphanRemoval = true)
     private Set<FaultReport> faultReports = new HashSet<>();
 
-    public int getTotalCount() {
-        return chargers.size();
+    public Long getTotalCount() {
+        return (long) chargers.size();
     }
 
-    public int getAvailableCount() {
-        return (int) chargers.stream()
+    public Long getAvailableCount() {
+        return chargers.stream()
                 .filter(Charger::isAvailable)
                 .count();
     }
